# specs/001-telco-adb-automation/contracts/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Telco ADB Automation API",
    "description": "REST API for managing devices, flows, and executions in the Telco ADB Automation Desktop Framework",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/devices/{device_id}": {
      "get": {
        "summary": "Get device details",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Device details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Device" }
              }
            }
          },
          "404": {
            "description": "Device not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Device": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "phone_number": { "type": "string", "nullable": true },
          "sim_info": { "type": "object" },
          "model": { "type": "string" },
          "os_version": { "type": "string" },
          "status": { "type": "string", "enum": ["connected", "disconnected", "busy", "error"] },
          "last_seen": { "type": "string", "format": "date-time" },
          "capabilities": { "type": "object" }
        }
      },
...
    }
  }
}